@@ -18,14 +18,15 @@ require 'spec_helper'
 describe BatchUpdateJob do
   before do
     @user = FactoryGirl.find_or_create(:user)
-    @batch = Batch.new
-    @batch.save
-    @file = GenericFile.new(:batch=>@batch)
-    @file.apply_depositor_metadata(@user.user_key)
-    @file.save
-    @file2 = GenericFile.new(:batch=>@batch)
-    @file2.apply_depositor_metadata('otherUser')
-    @file2.save
+    @batch = Batch.create
+    @file = GenericFile.new(batch: @batch).tap do |f|
+      f.apply_depositor_metadata(@user.user_key)
+      f.save
+    end
+    @file2 = GenericFile.new(batch: @batch).tap do |f|
+      f.apply_depositor_metadata('otherUser')
+      f.save
+    end
   end
   after do
     @batch.delete
@@ -34,8 +35,8 @@ describe BatchUpdateJob do
   end
   describe "failing update" do
     it "should check permissions for each file before updating" do
-      User.any_instance.expects(:can?).with(:edit, @file).returns(false)
-      User.any_instance.expects(:can?).with(:edit, @file2).returns(false)
+      User.any_instance.should_receive(:can?).with(:edit, @file).and_return(false)
+      User.any_instance.should_receive(:can?).with(:edit, @file2).and_return(false)
       params = {'generic_file' => {'read_groups_string' => '', 'read_users_string' => 'archivist1, archivist2', 'tag' => ['']}, 'id' => @batch.pid, 'controller' => 'batch', 'action' => 'update'}.with_indifferent_access
       BatchUpdateJob.new(@user.user_key, params).run
       @user.mailbox.inbox[0].messages[0].subject.should == "Batch upload permission denied"
@@ -44,14 +45,14 @@ describe BatchUpdateJob do
   end
   describe "passing update" do
     it "should log a content update event" do
-      User.any_instance.expects(:can?).with(:edit, @file).returns(true)
-      User.any_instance.expects(:can?).with(:edit, @file2).returns(true)
-      s1 = mock('one')
-      ContentUpdateEventJob.expects(:new).with(@file.pid, @user.user_key).returns(s1)
-      Sufia.queue.expects(:push).with(s1).once
-      s2 = mock('two')
-      ContentUpdateEventJob.expects(:new).with(@file2.pid, @user.user_key).returns(s2)
-      Sufia.queue.expects(:push).with(s2).once
+      User.any_instance.should_receive(:can?).with(:edit, @file).and_return(true)
+      User.any_instance.should_receive(:can?).with(:edit, @file2).and_return(true)
+      s1 = double('one')
+      ContentUpdateEventJob.should_receive(:new).with(@file.pid, @user.user_key).and_return(s1)
+      Sufia.queue.should_receive(:push).with(s1).once
+      s2 = double('two')
+      ContentUpdateEventJob.should_receive(:new).with(@file2.pid, @user.user_key).and_return(s2)
+      Sufia.queue.should_receive(:push).with(s2).once
       params = {'generic_file' => {'read_groups_string' => '', 'read_users_string' => 'archivist1, archivist2', 'tag' => ['“patience”']}, 'id' => @batch.pid, 'controller' => 'batch', 'action' => 'update'}.with_indifferent_access
       BatchUpdateJob.new(@user.user_key, params).run
       @user.mailbox.inbox[0].messages[0].subject.should == "Batch upload complete"
