@@ -20,6 +20,9 @@ class ApplicationController < ActionController::Base
   # Adds Sufia behaviors into the application controller
   include Sufia::Controller
 
+  # allow you to check for a valid user in the session
+  include Behaviors::HttpHeaderAuthenticatableBehavior
+
   # Please be sure to impelement current_user and user_session. Blacklight depends on
   # these methods in order to perform user specific actions.
 
@@ -95,29 +98,9 @@ class ApplicationController < ActionController::Base
     end
   end
 
-  def notifications_number
-    @notify_number = 0
-    @batches = []
-    return if action_name == "index" and controller_name == "mailbox"
-    if User.current
-      @notify_number = User.current.mailbox.inbox(:unread => true).count(:id, :distinct => true)
-      @batches = User.current.mailbox.inbox.map { |msg| msg.last_message.body[/<a class="batchid ui-helper-hidden">(.*)<\/a>The file(.*)/,1] }.select { |val| !val.blank? }
-    end
-  end
-
  protected
   def user_logged_in?
-    user_signed_in? and remote_user_set?
-  end
-
-  def remote_user_set?
-    return true if Rails.env.test?
-    # Unicorn seems to translate REMOTE_USER into HTTP_REMOTE_USER
-    if Rails.env.development?
-      request.env['HTTP_REMOTE_USER'].present?
-    else
-      request.env['REMOTE_USER'].present?
-    end
+    user_signed_in? and ( valid_user?(request.headers) || Rails.env.test?)
   end
 
   def has_access?
