@@ -70,36 +70,27 @@ class CatalogController < ApplicationController
   end
 
 
-  #####################
-  # jgm testing start #
-  #####################
-  if Rails.env == "integration"
-    # COPIED AND MODIFIED from:
-    #	/usr/local/rvm/gems/ree-1.8.7-2011.03@scholarsphere/gems/blacklight-3.3.2/lib/blacklight/catalog.rb
-    #
-    # when solr (RSolr) throws an error (RSolr::RequestError), this method is executed.
-    def rsolr_request_error(exception)
-      #if Rails.env == "development"
-      if ['development', 'integration'].include?(Rails.env)
-        raise exception # Rails own code will catch and give usual Rails error page with stack trace
+  # COPIED AND MODIFIED from:
+  #	/usr/local/rvm/gems/ree-1.8.7-2011.03@scholarsphere/gems/blacklight-3.3.2/lib/blacklight/catalog.rb
+  #
+  # when solr (RSolr) throws an error (RSolr::RequestError), this method is executed.
+  def rsolr_request_error(exception)
+    if ['development', 'test'].include?(Rails.env)
+      raise exception # Rails own code will catch and give usual Rails error page with stack trace
+    else
+      flash_notice = "Sorry, I don't understand your search."
+      # Set the notice flag if the flash[:notice] is already set to the error that we are setting.
+      # This is intended to stop the redirect loop error
+      notice = flash[:notice] if flash[:notice] == flash_notice
+      unless notice
+        flash[:notice] = flash_notice
+        redirect_to root_path, :status => 500
       else
-        flash_notice = "Sorry, I don't understand your search."
-        # Set the notice flag if the flash[:notice] is already set to the error that we are setting.
-        # This is intended to stop the redirect loop error
-        notice = flash[:notice] if flash[:notice] == flash_notice
-        unless notice
-          flash[:notice] = flash_notice
-          redirect_to root_path, :status => 500
-        else
-          render :template => "public/500.html", :layout => false, :status => 500
-        end
+        render :template => "public/500.html", :layout => false, :status => 500
       end
     end
-  #
   end
-  ###################
-  # jgm testing end #
-  ###################
+
 
   configure_blacklight do |config|
     ## Default parameters to send to solr for all search-like requests. See also SolrHelper#solr_search_params
@@ -384,6 +375,8 @@ class CatalogController < ApplicationController
     # If there are more than this many search results, no spelling ("did you
     # mean") suggestion is offered.
     config.spell_max = 5
+
+    config.http_method = :post
   end
 
   protected
