@@ -18,8 +18,7 @@ class GenericFile < ActiveFedora::Base
 
   has_file_datastream "full_text", versionable: false
 
-  delegate :proxy_depositor, :to=>:properties, :unique => true
-  delegate :on_behalf_of, :to=>:properties, :unique => true
+  has_attributes :proxy_depositor, :on_behalf_of, :datastream=>:properties, :multiple => false
 
   after_create :create_transfer_request
 
@@ -75,6 +74,22 @@ class GenericFile < ActiveFedora::Base
      end
      return stat
    end
+
+  def create_video_thumbnail
+    return unless Sufia.config.enable_ffmpeg
+
+    output_file = Dir::Tmpname.create(['sufia', ".png"], Sufia.config.temp_file_base){}
+    content.to_tempfile do |f|
+      # we could use something like this in order to find a frame in the middle.
+      #ffprobe -show_files video.avi 2> /dev/null | grep duration | cut -d= -f2 53.399999
+      command = "#{Sufia.config.ffmpeg_path} -i \"#{f.path}\" -loglevel quiet -vf \"scale=338:-1\"  -r  1  -t  1 -vframes 1 #{output_file}"
+      system(command)
+      raise "Unable to execute command \"#{command}\"" unless $?.success?
+    end
+
+    self.thumbnail.content = File.open(output_file, 'rb').read
+    self.thumbnail.mimeType = 'image/png'
+  end
 
 
   def extract_content
