@@ -14,7 +14,7 @@ require 'capistrano-notification'
 set :application, "scholarsphere"
 set :whenever_command, "bundle exec whenever"
 
-set :scm, :git 
+set :scm, :git
 set :deploy_via, :remote_cache
 set :repository,  "https://github.com/psu-stewardship/#{application}.git"
 
@@ -25,7 +25,7 @@ ssh_options[:keys] = [File.join(ENV["HOME"], ".ssh", "id_deploy_rsa")]
 set :use_sudo, false
 default_run_options[:pty] = true
 
-set :rbenv_ruby_version, "2.0.0-p247"
+set :rbenv_ruby_version, File.read(File.join(File.dirname(__FILE__), '..', '.ruby-version')).chomp
 set :rbenv_setup_shell, false
 
 notification.irc do |irc|
@@ -50,11 +50,14 @@ namespace :deploy do
   task :symlink_shared do
     run <<-CMD.compact
     ln -sf /dlt/#{application}/config_#{stage}/#{application}/database.yml #{release_path}/config/ &&
+    ln -sf /dlt/#{application}/config_#{stage}/#{application}/devise.yml #{release_path}/config/ &&
     ln -sf /dlt/#{application}/config_#{stage}/#{application}/fedora.yml #{release_path}/config/ &&
     ln -sf /dlt/#{application}/config_#{stage}/#{application}/hydra-ldap.yml #{release_path}/config/ &&
-    ln -sf /dlt/#{application}/config_#{stage}/#{application}/solr.yml #{release_path}/config/ &&
+    ln -sf /dlt/#{application}/config_#{stage}/#{application}/newrelic.yml #{release_path}/config/ &&
     ln -sf /dlt/#{application}/config_#{stage}/#{application}/redis.yml #{release_path}/config/ &&
-    ln -sf /dlt/#{application}/config_#{stage}/#{application}/secret_token.rb #{release_path}/config/initializers/
+    ln -sf /dlt/#{application}/config_#{stage}/#{application}/solr.yml #{release_path}/config/ &&
+    ln -sf /dlt/#{application}/config_#{stage}/#{application}/secret_token.rb #{release_path}/config/initializers/ &&
+    ln -sf /dlt/#{application}/config_#{stage}/#{application}/sufia-secret.rb #{release_path}/config/initializers/
     CMD
   end
 end
@@ -68,7 +71,7 @@ namespace :deploy do
   desc "Re-solrize objects"
   task :resolrize, :roles => :solr do
     run <<-CMD.compact
-    cd -- #{latest_release} && 
+    cd -- #{latest_release} &&
     RAILS_ENV=#{rails_env.to_s.shellescape} #{rake} #{application}:resolrize
     CMD
   end
