@@ -15,7 +15,7 @@ Sufia::Engine.configure do
   config.contact_email = 'DLT-GAMMA-PROJECT@lists.psu.edu'
   config.from_email = "ScholarSphere Form <scholarsphere-service-support@dlt.psu.edu>"
   config.logout_url = 'https://webaccess.psu.edu/cgi-bin/logout?http://localhost:3000/'
-  config.login_url = 'https://webaccess.psu.edu?cosign-localhost&http://localhost:3000/'
+  config.login_url = 'https://webaccess.psu.edu/?cosign-localhost&http://localhost:3000/'
 end
 ScholarSphere::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
@@ -26,8 +26,10 @@ ScholarSphere::Application.configure do
   # since you don't have to restart the webserver when you make code changes.
   config.cache_classes = false
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
+  # Do not eager load code on boot. This avoids loading your whole application
+  # just for the purpose of running a single test. If you are using a tool that
+  # preloads Rails for running tests, you may have to set it to true.
+  config.eager_load = false
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
