@@ -14,7 +14,7 @@
 
 class RedirectToWebAccessFailure < Devise::FailureApp
   def redirect_url
-    Sufia::Engine.config.login_url+ (request.env["ORIGINAL_FULLPATH"].blank? ? '' : request.env["ORIGINAL_FULLPATH"])
+    Sufia::Engine.config.login_url.chomp("/")+ (request.env["ORIGINAL_FULLPATH"].blank? ? '' : request.env["ORIGINAL_FULLPATH"])
   end
 
   def respond
@@ -23,5 +23,15 @@ class RedirectToWebAccessFailure < Devise::FailureApp
     else
       redirect
     end
+  end
+
+  # Overriding, so that we don't set the flash[:alert] with the unauthenticated message
+  def redirect
+    store_location!
+    if flash[:timedout] && flash[:alert]
+      flash.keep(:timedout)
+      flash.keep(:alert)
+    end
+    redirect_to redirect_url
   end
 end
