@@ -15,6 +15,7 @@
 require 'spec_helper'
 
 describe AuthoritiesController do
+  routes { Sufia::Engine.routes }
   describe "#query" do
     it "should return an array of hashes" do
       mock_hits = [{:label => "English", :uri => "http://example.org/eng"}, 
@@ -23,7 +24,7 @@ describe AuthoritiesController do
                    {:label => "Edgar", :uri => "http://example.org/edga"}, 
                    {:label => "Eddie", :uri => "http://example.org/edd"},
                    {:label => "Economics", :uri => "http://example.org/eco"}]
-      LocalAuthority.expects(:entries_by_term).returns(mock_hits)
+      LocalAuthority.should_receive(:entries_by_term).and_return(mock_hits)
       xhr :get, :query, :model=>"generic_files", :term=>"subject", :q=>"E"
       response.should be_success
       JSON.parse(response.body).count.should == 6
