@@ -17,11 +17,17 @@ require 'spec_helper'
 describe CollectionsController do
   before(:each) { @routes = Hydra::Collections::Engine.routes }
   before do
-    controller.stubs(:has_access?).returns(true)
+    controller.stub(:has_access?).and_return(true)
 
     @user = FactoryGirl.find_or_create(:user)
     sign_in @user
-    User.any_instance.stubs(:groups).returns([])
+    User.any_instance.stub(:groups).and_return([])
+  end
+
+  after (:all) do
+    Collection.destroy_all
+    GenericFile.destroy_all
+    User.destroy_all
   end
 
   describe '#new' do
@@ -30,10 +36,10 @@ describe CollectionsController do
       assigns(:collection).should be_kind_of(Collection)
     end
   end
-  
+
   describe '#create' do
     it "should create a Collection" do
-      controller.expects(:has_access?).returns(true)
+      controller.should_receive(:has_access?).and_return(true)
       old_count = Collection.count
       post :create, collection: {title: "My First Collection ", description: "The Description\r\n\r\nand more"}
       Collection.count.should == old_count+1
@@ -48,7 +54,7 @@ describe CollectionsController do
       @asset3 = GenericFile.new(title: "Third of the Assets", depositor:'abc')
       @asset3.apply_depositor_metadata('abc')
       @asset3.save
-      controller.expects(:has_access?).returns(true)
+      controller.should_receive(:has_access?).and_return(true)
       old_count = Collection.count
       post :create, collection: {title: "My own Collection ", description: "The Description\r\n\r\nand more"}, batch_document_ids:[@asset1.id, @asset2.id, @asset3.id]
       Collection.count.should == old_count+1
@@ -79,7 +85,7 @@ describe CollectionsController do
 
   describe "#update" do
     before do
-      @collection = Collection.new
+      @collection = Collection.new(title: "Collection Title")
       @collection.apply_depositor_metadata(@user.user_key)
       @collection.save
       @asset1 = GenericFile.new(title: "First of the Assets")
@@ -139,8 +145,8 @@ describe CollectionsController do
       @collection.apply_depositor_metadata(@user.user_key)
       @collection.members = [@asset1,@asset2,@asset3]
       @collection.save
-      controller.stubs(:authorize!).returns(true)
-      controller.stubs(:apply_gated_search)
+      controller.stub(:authorize!).and_return(true)
+      controller.stub(:apply_gated_search)
     end
     it "should return the collection and its members" do
       get :show, id: @collection.id
@@ -151,14 +157,14 @@ describe CollectionsController do
       ids.should include @asset3.pid
       ids.should_not include @asset4.pid
     end
-    it "should query the collection members" do
-      pending "The query isn't working here for some reason.  This is covered by a test in features/collection_spec.rb"
-      get :show, id: @collection.id, cq:@asset1.title, id: @collection.pid
-      assigns[:collection].title.should == @collection.title
-      ids = assigns[:member_docs].map {|d| d.id}
-      ids.should include @asset1.pid
-      ids.should_not include @asset2.pid
-      ids.should_not include @asset3.pid
+    context "signed out" do
+      before do
+        sign_out @user
+      end
+      it "should not show me files in the collection" do
+        get :show, id: @collection.id
+        assigns[:member_docs].count.should == 0
+      end
     end
   end
 end
