@@ -15,33 +15,34 @@
 require 'spec_helper'
 
 describe DashboardController do
+  routes { Sufia::Engine.routes }
   before do
-    User.any_instance.stubs(:groups).returns([])
+    User.any_instance.stub(:groups).and_return([])
   end
   # This doesn't really belong here, but it works for now
   describe "authenticate!" do
     before(:each) do
       @user = FactoryGirl.find_or_create(:archivist)
-      request.stubs(:headers).returns('REMOTE_USER' => @user.login).at_least_once
+      request.stub(:headers).and_return('REMOTE_USER' => @user.login)
       @strategy = Devise::Strategies::HttpHeaderAuthenticatable.new(nil)
-      @strategy.expects(:request).returns(request).at_least_once
+      @strategy.stub(:request => request)
     end
     after(:each) do
       @user.delete
     end
     it "should populate LDAP attrs if user is new" do
-      User.stubs(:find_by_login).with(@user.login).returns(nil)
-      User.expects(:create).with(login: @user.login).returns(@user).once
-      User.any_instance.expects(:populate_attributes).once
+      User.stub(:find_by_login).with(@user.login).and_return(nil)
+      User.should_receive(:create).with(login: @user.login, email:@user.login).once.and_return(@user)
+      User.any_instance.should_receive(:populate_attributes).once
       @strategy.should be_valid
       @strategy.authenticate!.should == :success
       sign_in @user
       get :index
     end
     it "should not populate LDAP attrs if user is not new" do
-      User.stubs(:find_by_login).with(@user.login).returns(@user)
-      User.expects(:create).with(login: @user.login).never
-      User.any_instance.expects(:populate_attributes).never
+      User.stub(:find_by_login).with(@user.login).and_return(@user)
+      User.should_receive(:create).with(login: @user.login).never
+      User.any_instance.should_receive(:populate_attributes).never
       @strategy.should be_valid
       @strategy.authenticate!.should == :success
       sign_in @user
@@ -52,7 +53,7 @@ describe DashboardController do
     before (:each) do
       @user = FactoryGirl.find_or_create(:archivist)
       sign_in @user
-      User.any_instance.stubs(:groups).returns([])
+      User.any_instance.stub(:groups).and_return([])
     end
     describe "#index" do
       before (:each) do
