@@ -1,4 +1,3 @@
-
 # Copyright © 2012 The Pennsylvania State University
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -21,7 +20,7 @@ describe UsersController do
     @user = FactoryGirl.find_or_create(:user)
     @another_user = FactoryGirl.find_or_create(:archivist)
     sign_in @user
-    User.any_instance.stubs(:groups).returns([])
+    User.any_instance.stub(:groups).and_return([])
   end
   after(:all) do
     @user = FactoryGirl.find(:user) rescue
@@ -31,13 +30,13 @@ describe UsersController do
   end
   describe "#show" do
     it "show the user profile if user exists" do
-      get :show, uid: @user.login
+      get :show, id: @user.login
       response.should be_success
       response.should_not redirect_to(root_path)
       flash[:alert].should be_nil
     end
     it "redirects to root if user does not exist" do
-      get :show, uid: 'johndoe666'
+      get :show, id: 'johndoe666'
       response.should redirect_to(root_path)
       flash[:alert].should include ("User 'johndoe666' does not exist")
     end
@@ -45,9 +44,9 @@ describe UsersController do
       @file = GenericFile.new
       @file.apply_depositor_metadata(@user.login)
       @file.save
-      post :toggle_trophy, {uid: @user.login, file_id: @file.pid["scholarsphere:".length..-1]}
+      post :toggle_trophy, {id: @user.login, file_id: @file.pid["scholarsphere:".length..-1]}
       @file.delete
-      get :show, uid: @user.login
+      get :show, id: @user.login
       response.should be_success
       response.should_not redirect_to(root_path)
       flash[:alert].should be_nil
@@ -69,20 +68,20 @@ describe UsersController do
       @user3 = FactoryGirl.find_or_create(:archivist)
       get :index, format: :json
       json = JSON.parse(response.body)
-      json.map{|u| u['id'] }.should include(@user.login, @user2.login, @user3.login)
-      json.map{|u| u['text'] }.should include(@user.login, @user2.login, @user3.login)
+      json.map{|u| u['id'] }.should match_array([@user.login, @user2.login, @user3.login])
+      json.map{|u| u['text'] }.should match_array(["#{@user.display_name} (#{@user.to_s})", @user2.to_s, @user3.to_s])
       end
     end
   end
   describe "#edit" do
     it "show edit form when user edits own profile" do
-      get :edit, uid: @user.login
+      get :edit, id: @user.login
       response.should be_success
       response.should render_template('users/edit')
       flash[:alert].should be_nil
     end
     it "redirects to show profile when user attempts to edit another profile" do
-      get :edit, uid: @another_user.login
+      get :edit, id: @another_user.login
       response.should redirect_to(@routes.url_helpers.profile_path(@another_user.login))
       flash[:alert].should include("Permission denied: cannot access this page.")
     end
@@ -90,9 +89,9 @@ describe UsersController do
       @file = GenericFile.new
       @file.apply_depositor_metadata(@user.login)
       @file.save
-      post :toggle_trophy, {uid: @user.login, file_id: @file.pid["scholarsphere:".length..-1]}
+      post :toggle_trophy, {id: @user.login, file_id: @file.pid["scholarsphere:".length..-1]}
       @file.delete
-      get :show, uid: @user.login
+      get :show, id: @user.login
       response.should be_success
       response.should_not redirect_to(root_path)
       flash[:alert].should be_nil
@@ -100,55 +99,50 @@ describe UsersController do
   end
   describe "#update" do
     it "should not allow other users to update" do
-      post :update, uid: @another_user.login, user: { avatar: nil }
+      post :update, id: @another_user.login, user: { avatar: nil }
       response.should redirect_to(@routes.url_helpers.profile_path(@another_user.login))
       flash[:alert].should include("Permission denied: cannot access this page.")
     end
     it "should set an avatar and redirect to profile" do
-      @user.avatar.file?.should be_false
-      s1 = mock('one')
-      UserEditProfileEventJob.expects(:new).with(@user.login).returns(s1)
-      Sufia.queue.expects(:push).with(s1).once
-      #Resque.expects(:enqueue).with(UserEditProfileEventJob, @user.login).once
+      @user.avatar?.should be_false
+      s1 = double('one')
+      UserEditProfileEventJob.should_receive(:new).with(@user.login).and_return(s1)
+      Sufia.queue.should_receive(:push).with(s1).once
       f = fixture_file_upload('/world.png', 'image/png')
-      post :update, uid: @user.login, user: { avatar: f }
+      post :update, id: @user.login, user: { avatar: f }
       response.should redirect_to(@routes.url_helpers.profile_path(@user.login))
       flash[:notice].should include("Your profile has been updated")
-      User.find_by_login(@user.login).avatar.file?.should be_true
+      User.find_by_login(@user.login).avatar?.should be_true
     end
     it "should validate the content type of an avatar" do
-      UserEditProfileEventJob.expects(:new).with(@user.login).never
-      #Resque.expects(:enqueue).with(UserEditProfileEventJob, @user.login).never
+      UserEditProfileEventJob.should_receive(:new).with(@user.login).never
       f = fixture_file_upload('/image.jp2', 'image/jp2')
-      post :update, uid: @user.login, user: { avatar: f }
+      post :update, id: @user.login, user: { avatar: f }
       response.should redirect_to(@routes.url_helpers.edit_profile_path(@user.login))
-      flash[:alert].should include("Avatar content type is invalid")
+      flash[:alert].should include("Avatar You are not allowed to upload \"jp2\" files, allowed types: jpg, jpeg, png, gif, bmp, tif, tiff")
     end
     it "should validate the size of an avatar" do
       f = fixture_file_upload('/4-20.png', 'image/png')
-      UserEditProfileEventJob.expects(:new).with(@user.login).never
-      #Resque.expects(:enqueue).with(UserEditProfileEventJob, @user.login).never
-      post :update, uid: @user.login, user: { avatar: f }
+      UserEditProfileEventJob.should_receive(:new).with(@user.login).never
+      post :update, id: @user.login, user: { avatar: f }
       response.should redirect_to(@routes.url_helpers.edit_profile_path(@user.login))
-      flash[:alert].should include("Avatar file size must be less than 2097152 Bytes")
+      flash[:alert].should include("Avatar file size must be less than 2MB")
     end
     it "should delete an avatar" do
-      s1 = mock('one')
-      UserEditProfileEventJob.expects(:new).with(@user.login).returns(s1)
-      Sufia.queue.expects(:push).with(s1).once
-      #Resque.expects(:enqueue).with(UserEditProfileEventJob, @user.login).once
-      post :update, uid: @user.login, delete_avatar: true
+      s1 = double('one')
+      UserEditProfileEventJob.should_receive(:new).with(@user.login).and_return(s1)
+      Sufia.queue.should_receive(:push).with(s1).once
+      post :update, id: @user.login, delete_avatar: true
       response.should redirect_to(@routes.url_helpers.profile_path(@user.login))
       flash[:notice].should include("Your profile has been updated")
-      @user.avatar.file?.should be_false
+      @user.avatar?.should be_false
     end
     it "should refresh directory attributes" do
-      s1 = mock('one')
-      UserEditProfileEventJob.expects(:new).with(@user.login).returns(s1)
-      Sufia.queue.expects(:push).with(s1).once
-      #Resque.expects(:enqueue).with(UserEditProfileEventJob, @user.login).once
-      User.any_instance.expects(:populate_attributes).once
-      post :update, uid: @user.login, update_directory: true
+      s1 = double('one')
+      UserEditProfileEventJob.should_receive(:new).with(@user.login).and_return(s1)
+      Sufia.queue.should_receive(:push).with(s1).once
+      User.any_instance.should_receive(:populate_attributes).once
+      post :update, id: @user.login, update_directory: true
       response.should redirect_to(@routes.url_helpers.profile_path(@user.login))
       flash[:notice].should include("Your profile has been updated")
     end
@@ -156,7 +150,7 @@ describe UsersController do
       @user.twitter_handle.blank?.should be_true
       @user.facebook_handle.blank?.should be_true
       @user.googleplus_handle.blank?.should be_true
-      post :update, uid: @user.login, user: { twitter_handle: 'twit', facebook_handle: 'face', googleplus_handle: 'goo' }
+      post :update, id: @user.login, user: { twitter_handle: 'twit', facebook_handle: 'face', googleplus_handle: 'goo' }
       response.should redirect_to(@routes.url_helpers.profile_path(@user.login))
       flash[:notice].should include("Your profile has been updated")
       u = User.find_by_login(@user.login)
@@ -171,53 +165,47 @@ describe UsersController do
     end
     it "should follow another user if not already following, and log an event" do
       @user.following?(@another_user).should be_false
-      s1 = mock('one')
-      UserFollowEventJob.expects(:new).with(@user.login, @another_user.login).returns(s1)
-      Sufia.queue.expects(:push).with(s1).once
-      #Resque.expects(:enqueue).with(UserFollowEventJob, @user.login, @another_user.login).once
-      post :follow, uid: @another_user.login
+      s1 = double('one')
+      UserFollowEventJob.should_receive(:new).with(@user.login, @another_user.login).and_return(s1)
+      Sufia.queue.should_receive(:push).with(s1).once
+      post :follow, id: @another_user.login
       response.should redirect_to(@routes.url_helpers.profile_path(@another_user.login))
       flash[:notice].should include("You are following #{@another_user.login}")
     end
     it "should redirect to profile if already following and not log an event" do
-      User.any_instance.stubs(:following?).with(@another_user).returns(true)
-      UserFollowEventJob.expects(:new).with(@user.login, @another_user.login).never
-      #Resque.expects(:enqueue).with(UserFollowEventJob, @user.login, @another_user.login).never
-      post :follow, uid: @another_user.login
+      User.any_instance.stub(:following?).with(@another_user).and_return(true)
+      UserFollowEventJob.should_receive(:new).with(@user.login, @another_user.login).never
+      post :follow, id: @another_user.login
       response.should redirect_to(@routes.url_helpers.profile_path(@another_user.login))
       flash[:notice].should include("You are following #{@another_user.login}")
     end
     it "should redirect to profile if user attempts to self-follow and not log an event" do
-      UserFollowEventJob.expects(:new).with(@user.login, @another_user.login).never
-      #Resque.expects(:enqueue).with(UserFollowEventJob, @user.login, @user.login).never
-      post :follow, uid: @user.login
+      UserFollowEventJob.should_receive(:new).with(@user.login, @another_user.login).never
+      post :follow, id: @user.login
       response.should redirect_to(@routes.url_helpers.profile_path(@user.login))
       flash[:alert].should include("You cannot follow or unfollow yourself")
     end
   end
   describe "#unfollow" do
     it "should unfollow another user if already following, and log an event" do
-      User.any_instance.stubs(:following?).with(@another_user).returns(true)
-      s1 = mock('one')
-      UserUnfollowEventJob.expects(:new).with(@user.login, @another_user.login).returns(s1)
-      Sufia.queue.expects(:push).with(s1).once
-      #Resque.expects(:enqueue).with(UserUnfollowEventJob, @user.login, @another_user.login).once
-      post :unfollow, uid: @another_user.login
+      User.any_instance.stub(:following?).with(@another_user).and_return(true)
+      s1 = double('one')
+      UserUnfollowEventJob.should_receive(:new).with(@user.login, @another_user.login).and_return(s1)
+      Sufia.queue.should_receive(:push).with(s1).once
+      post :unfollow, id: @another_user.login
       response.should redirect_to(@routes.url_helpers.profile_path(@another_user.login))
       flash[:notice].should include("You are no longer following #{@another_user.login}")
     end
     it "should redirect to profile if not following and not log an event" do
-      @user.stubs(:following?).with(@another_user).returns(false)
-      UserFollowEventJob.expects(:new).with(@user.login, @another_user.login).never
-      #Resque.expects(:enqueue).with(UserUnfollowEventJob, @user.login, @another_user.login).never
-      post :unfollow, uid: @another_user.login
+      @user.stub(:following?).with(@another_user).and_return(false)
+      UserFollowEventJob.should_receive(:new).with(@user.login, @another_user.login).never
+      post :unfollow, id: @another_user.login
       response.should redirect_to(@routes.url_helpers.profile_path(@another_user.login))
       flash[:notice].should include("You are no longer following #{@another_user.login}")
     end
     it "should redirect to profile if user attempts to self-follow and not log an event" do
-      UserFollowEventJob.expects(:new).with(@user.login, @another_user.login).never
-      #Resque.expects(:enqueue).with(UserUnfollowEventJob, @user.login, @user.login).never
-      post :unfollow, uid: @user.login
+      UserFollowEventJob.should_receive(:new).with(@user.login, @another_user.login).never
+      post :unfollow, id: @user.login
       response.should redirect_to(@routes.url_helpers.profile_path(@user.login))
       flash[:alert].should include("You cannot follow or unfollow yourself")
     end
@@ -232,18 +220,18 @@ describe UsersController do
        @file.delete
      end
      it "should trophy a file" do
-      post :toggle_trophy, {uid: @user.login, file_id: @file.pid["scholarsphere:".length..-1]}
-      JSON.parse(response.body)['trophy']['user_id'].should == @user.id
-      JSON.parse(response.body)['trophy']['generic_file_id'].should == @file.pid["scholarsphere:".length..-1]
+      post :toggle_trophy, {id: @user.login, file_id: @file.pid["scholarsphere:".length..-1]}
+      JSON.parse(response.body)['user_id'].should == @user.id
+      JSON.parse(response.body)['generic_file_id'].should == @file.pid["scholarsphere:".length..-1]
     end
      it "should not trophy a file for a different user" do
-      post :toggle_trophy, {uid: @another_user.login, file_id: @file.pid}
+      post :toggle_trophy, {id: @another_user.login, file_id: @file.pid}
       response.should_not be_success
     end
      it "should not trophy a file with no edit privs" do
       sign_out @user
       sign_in @another_user
-      post :toggle_trophy, {uid: @another_user.login, file_id: @file.pid}
+      post :toggle_trophy, {id: @another_user.login, file_id: @file.pid}
       response.should_not be_success
     end
   end
