@@ -28,5 +28,8 @@ describe Collection do
     @collection.save
     (@collection.datastreams["rightsMetadata"].permissions({group:"public"})).should == "read"
   end
-
+  it "should not allow a collection to be saved without a title" do
+     @collection.title = nil
+     expect{@collection.save!}.to raise_error(ActiveFedora::RecordInvalid)
+  end
 end
