@@ -14,40 +14,54 @@
 
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV["RAILS_ENV"] ||= 'test'
+require 'rake'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
 require 'rspec/autorun'
 require 'capybara/rspec'
 require 'capybara/rails'
-require 'mocha/setup'
+require 'capybara/poltergeist'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
-Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
+Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
 
 RSpec.configure do |config|
   config.before(:suite) do
     before_files_count = GenericFile.count
     before_batches_count = Batch.count
-    puts "WARNING: Your jetty is not clean, so tests may be funky! (#{before_files_count} files, #{before_batches_count} batches)" if before_files_count > 0 or before_batches_count > 0
+    before_collections_count = Collection.count
+    puts "WARNING: Your jetty is not clean, so tests may be funky! (#{before_files_count} files, #{before_batches_count} batches, #{before_collections_count} collections)" if before_files_count > 0 or before_batches_count > 0 or before_collections_count > 0
   end
+
   config.after(:all) do
     files_count = GenericFile.count
     batches_count = Batch.count
+    collections_count = Collection.count
     puts "WARNING: #{files_count} files need cleaning up" if files_count > 0
     puts "WARNING: #{batches_count} batches need cleaning up" if batches_count > 0
+    puts "WARNING: #{collections_count} batches need cleaning up" if collections_count > 0
   end
 
-  Capybara.register_driver :selenium do |app|
-    profile = Selenium::WebDriver::Firefox::Profile.new
-    Capybara::Selenium::Driver.new(app, :profile => profile)
+  config.after(:each) do
+    # Unspoof the warden strategy.  See spec/support/user_login.rb for more info.
+    unspoof_http_auth if example.options[:js]
   end
 
-  # == Mock Framework
-  #
-  # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
-  #
-  config.mock_with :mocha
+  # Make Capybara wait a bit longer so sluggish AJAX reqs can finish
+  # on Travis
+  Capybara.default_wait_time = 15
+
+  # Use poltergeist (phantomjs driver) for feature tests requiring javascript
+  Capybara.register_driver :poltergeist do |app|
+    Capybara::Poltergeist::Driver.new(app, {
+        js_errors: false,
+        inspector: true,
+        phantomjs_options: ['--proxy-type=none']
+      })
+  end
+  Capybara.default_driver = :poltergeist
+  Capybara.javascript_driver = :poltergeist
 
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   config.fixture_path = "#{::Rails.root}/spec/fixtures"
